+++ travis/before_install.sh
@@ -1,2 +1,0 @@
-if [ ! "$NOXDEBUG" = "yes" ]; then phpenv config-rm xdebug.ini; fi;
-composer self-update