+++ travis/install_checks.sh
@@ -1,3 +1,0 @@
-composer install
-composer require satooshi/php-coveralls --dev;
-composer require squizlabs/php_codesniffer="~2.3" --dev;